new wots.mjs
This is modular javascript, if you don't want it just rename to .js and remove the last line "export [..]".
The code is not dependent. You can of course make more ordered wrappers. The code is self explaining but if you need help don't hesitate contacting me on Discord NickP05#6940 . To see examples look at example.mjs